requirements.txt
python-dotenv>=1.0.0
azure-cosmos>=4.3.0
loguru>=0.7.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
